requirements/app.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/app.txt requirements/app.in
#
alembic==1.14.0
    # via flask-migrate
altair==5.5.0
    # via flexmeasures
argon2-cffi==23.1.0
    # via flexmeasures
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via rq-dashboard
attrs==24.3.0
    # via
    #   jsonschema
    #   referencing
babel==2.16.0
    # via py-moneyed
bcrypt==4.0.1
    # via flexmeasures
blinker==1.9.0
    # via
    #   flask
    #   flask-mail
    #   flask-principal
    #   sentry-sdk
certifi==2024.12.14
    # via
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   click-default-group
    #   flask
    #   flexmeasures
    #   rq
    #   s2-python
click-default-group==1.2.4
    # via flexmeasures
contourpy==1.3.1
    # via matplotlib
convertdate==2.4.0
    # via workalendar
cycler==0.12.1
    # via matplotlib
dill==0.3.9
    # via openturns
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   flask-security-too
    #   flexmeasures
filelock==3.16.1
    # via tldextract
flask==3.1.0
    # via
    #   flask-classful
    #   flask-cors
    #   flask-json
    #   flask-login
    #   flask-mail
    #   flask-marshmallow
    #   flask-migrate
    #   flask-principal
    #   flask-security-too
    #   flask-sqlalchemy
    #   flask-sslify
    #   flask-wtf
    #   flexmeasures
    #   rq-dashboard
    #   sentry-sdk
flask-classful==0.16.0
    # via flexmeasures
flask-cors==5.0.0
    # via flexmeasures
flask-json==0.4.0
    # via flexmeasures
flask-login==0.6.3
    # via
    #   flask-security-too
    #   flexmeasures
flask-mail==0.10.0
    # via flexmeasures
flask-marshmallow==1.3.0
    # via flexmeasures
flask-migrate==4.1.0
    # via flexmeasures
flask-principal==0.4.0
    # via flask-security-too
flask-security-too==5.5.2
    # via flexmeasures
flask-sqlalchemy==3.1.1
    # via
    #   flask-migrate
    #   flexmeasures
flask-sslify==0.1.5
    # via flexmeasures
flask-wtf==1.2.2
    # via
    #   flask-security-too
    #   flexmeasures
flexcache==0.3
    # via pint
flexmeasures==0.24.0
    # via -r requirements/app.in
flexparser==0.4
    # via pint
fonttools==4.55.3
    # via matplotlib
greenlet==3.1.1
    # via sqlalchemy
humanize==4.11.0
    # via flexmeasures
idna==3.10
    # via
    #   email-validator
    #   requests
    #   tldextract
importlib-metadata==8.5.0
    # via
    #   flexmeasures
    #   timely-beliefs
importlib-resources==6.5.2
    # via flask-security-too
inflect==6.0.2
    # via flexmeasures
inflection==0.5.1
    # via flexmeasures
iso8601==2.1.0
    # via flexmeasures
isodate==0.7.2
    # via
    #   flexmeasures
    #   timely-beliefs
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.5
    # via
    #   altair
    #   flask
joblib==1.4.2
    # via
    #   scikit-learn
    #   sktime
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
kiwisolver==1.4.8
    # via matplotlib
lunardate==0.2.2
    # via workalendar
mako==1.3.8
    # via alembic
markupsafe==3.0.2
    # via
    #   flask-security-too
    #   jinja2
    #   mako
    #   sentry-sdk
    #   werkzeug
    #   wtforms
marshmallow==3.25.1
    # via
    #   flask-marshmallow
    #   flexmeasures
    #   marshmallow-polyfield
    #   marshmallow-sqlalchemy
    #   webargs
marshmallow-polyfield==5.11
    # via flexmeasures
marshmallow-sqlalchemy==1.3.0
    # via flexmeasures
matplotlib==3.10.0
    # via timetomodel
narwhals==1.22.0
    # via altair
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   patsy
    #   properscoring
    #   scikit-learn
    #   scipy
    #   sktime
    #   statsmodels
    #   timely-beliefs
    #   timetomodel
    #   uniplot
openturns==1.24
    # via timely-beliefs
packaging==24.2
    # via
    #   altair
    #   marshmallow
    #   matplotlib
    #   sktime
    #   statsmodels
    #   webargs
pandas==2.2.1
    # via
    #   flexmeasures
    #   sktime
    #   statsmodels
    #   timely-beliefs
    #   timetomodel
passlib==1.7.4
    # via flask-security-too
patsy==1.0.1
    # via statsmodels
pillow==11.1.0
    # via
    #   flexmeasures
    #   matplotlib
pint==0.24.4
    # via flexmeasures
platformdirs==4.3.6
    # via pint
ply==3.11
    # via pyomo
properscoring==0.1
    # via timely-beliefs
psutil==6.1.1
    # via openturns
psycopg2-binary==2.9.10
    # via
    #   flexmeasures
    #   timely-beliefs
py-moneyed==3.0
    # via flexmeasures
pycparser==2.22
    # via cffi
pydantic==1.10.20
    # via
    #   flexmeasures
    #   inflect
    #   s2-python
pyluach==2.2.0
    # via workalendar
pymeeus==0.5.12
    # via convertdate
pyomo==6.8.2
    # via flexmeasures
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   matplotlib
    #   pandas
    #   timetomodel
    #   workalendar
python-dotenv==1.0.1
    # via flexmeasures
pytz==2024.2
    # via
    #   flexmeasures
    #   pandas
    #   s2-python
    #   timely-beliefs
    #   timetomodel
pyyaml==6.0.2
    # via flexmeasures
redis==5.2.1
    # via
    #   flexmeasures
    #   redis-sentinel-url
    #   rq
    #   rq-dashboard
redis-sentinel-url==1.0.1
    # via rq-dashboard
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   requests-file
    #   tldextract
requests-file==2.1.0
    # via tldextract
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
rq==2.1.0
    # via
    #   flexmeasures
    #   rq-dashboard
rq-dashboard==0.8.2.2
    # via flexmeasures
s2-python==0.1.3
    # via -r requirements/app.in
scikit-base==0.12.0
    # via sktime
scikit-learn==1.5.2
    # via
    #   sktime
    #   timetomodel
scipy==1.15.1
    # via
    #   properscoring
    #   scikit-learn
    #   sktime
    #   statsmodels
    #   timely-beliefs
    #   timetomodel
sentry-sdk[flask]==2.19.2
    # via flexmeasures
six==1.17.0
    # via python-dateutil
sktime==0.35.0
    # via timely-beliefs
sqlalchemy==2.0.37
    # via
    #   alembic
    #   flask-sqlalchemy
    #   flexmeasures
    #   marshmallow-sqlalchemy
    #   timely-beliefs
    #   timetomodel
statsmodels==0.14.4
    # via timetomodel
tabulate==0.9.0
    # via flexmeasures
threadpoolctl==3.5.0
    # via scikit-learn
timely-beliefs[forecast]==3.1.0
    # via flexmeasures
timetomodel==0.7.3
    # via flexmeasures
tldextract==5.1.3
    # via flexmeasures
types-python-dateutil==2.9.0.20241206
    # via arrow
typing-extensions==4.12.2
    # via
    #   alembic
    #   altair
    #   flexcache
    #   flexparser
    #   pint
    #   py-moneyed
    #   pydantic
    #   sqlalchemy
tzdata==2024.2
    # via pandas
uniplot==0.16.2
    # via flexmeasures
urllib3==2.3.0
    # via
    #   requests
    #   sentry-sdk
vl-convert-python==1.7.0
    # via flexmeasures
webargs==8.6.0
    # via flexmeasures
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
    #   flexmeasures
workalendar==17.0.0
    # via flexmeasures
wtforms==3.2.1
    # via
    #   flask-security-too
    #   flask-wtf
xlrd==2.0.1
    # via flexmeasures
zipp==3.21.0
    # via importlib-metadata
